Check the capabilities of libjpeg
It looks like libjpeg is capable of extracting nearly all information
needed.
But it is also important to note that the image is decoded. And
strictly speaking that's not necessary. On the other hand it should also
be said that the library is quite fast and therefore the decoding should
not become a problem.