Include functionallity for printing metadata
Previous to this change the default behaviour was to check for
compliance with RFC 2435. But it also is required to show the actual
values.
Therefore the default behaviour is changed to print the metadata instead
of checking for compliance. The compliance check gets associated with
the option "-c" or "--compliance".